Test for existing file before trying to remove

+++ framework/includes/Search/Results/Cache/Controller.php
@@ -66,6 +66,11 @@ class Controller implements CacheInterface
 	{
 		$cacheFile = $this->_getCacheFilePath( $key );
 
+		// Check the cache file exists
+		if ( !is_readable( $cacheFile ) ) {
+			return false;
+		}
+
 		unlink( $cacheFile );
 
 		return $this;
